Move components out of shared and into their own dirs
