+++ projects/js-packages/charts/src/components/chart-composition/chart-svg.tsx
