+++ projects/js-packages/charts/src/components/chart-composition/types.ts
@@ -1,4 +1,4 @@
-import type { Legend } from '../../legend';
+import type { Legend } from '../legend';
 import type { ComponentType, FC, PropsWithChildren } from 'react';
 
 /**
