+++ projects/js-packages/charts/src/components/default-glyph/default-glyph.tsx
