+++ projects/js-packages/charts/src/components/default-glyph/index.ts
@@ -1,0 +1,1 @@
+export { DefaultGlyph } from './default-glyph';
