+++ projects/js-packages/charts/src/components/line-chart/line-chart.tsx
@@ -20,17 +20,17 @@ import {
 	useGlobalChartsContext,
 } from '../../providers/chart-context';
 import { attachSubComponents, getSeriesLineStyles } from '../../utils';
+import { DefaultGlyph } from '../default-glyph';
 import { Legend } from '../legend';
 import { useChartLegendData } from '../legend/use-chart-legend-data';
-import { DefaultGlyph } from '../shared/default-glyph';
-import { SingleChartContext, type SingleChartRef } from '../shared/single-chart-context';
-import { withResponsive } from '../shared/with-responsive';
+import { SingleChartContext, type SingleChartRef } from '../single-chart-context';
 import { AccessibleTooltip, useKeyboardNavigation } from '../tooltip/accessible-tooltip';
+import { withResponsive } from '../with-responsive';
 import LineChartAnnotation from './line-chart-annotation';
 import LineChartAnnotationsOverlay from './line-chart-annotations-overlay';
 import styles from './line-chart.module.scss';
 import type { BaseChartProps, DataPoint, DataPointDate, SeriesData, Optional } from '../../types';
-import type { ResponsiveConfig } from '../shared/with-responsive';
+import type { ResponsiveConfig } from '../with-responsive';
 import type { TickFormatter } from '@visx/axis';
 import type { GlyphProps } from '@visx/xychart';
 import type { RenderTooltipParams } from '@visx/xychart/lib/components/Tooltip';
