+++ projects/js-packages/charts/src/components/pie-semi-circle-chart/pie-semi-circle-chart.tsx
@@ -14,16 +14,16 @@ import {
 	GlobalChartsContext,
 } from '../../providers/chart-context';
 import { attachSubComponents } from '../../utils';
+import { ChartSVG, ChartHTML, useChartChildren } from '../chart-composition';
 import { Legend } from '../legend';
 import { useChartLegendData } from '../legend/use-chart-legend-data';
-import { ChartSVG, ChartHTML, useChartChildren } from '../shared/chart-composition';
-import { SingleChartContext } from '../shared/single-chart-context';
-import { withResponsive } from '../shared/with-responsive';
+import { SingleChartContext } from '../single-chart-context';
 import { BaseTooltip } from '../tooltip';
+import { withResponsive } from '../with-responsive';
 import styles from './pie-semi-circle-chart.module.scss';
 import type { BaseChartProps, DataPointPercentage, Optional } from '../../types';
-import type { ChartComponentWithComposition } from '../shared/chart-composition';
-import type { ResponsiveConfig } from '../shared/with-responsive';
+import type { ChartComponentWithComposition } from '../chart-composition';
+import type { ResponsiveConfig } from '../with-responsive';
 import type { PieArcDatum } from '@visx/shape/lib/shapes/Pie';
 import type { FC, MouseEvent, ReactNode } from 'react';
