+++ projects/js-packages/charts/src/components/with-responsive/index.ts
@@ -1,0 +1,2 @@
+export { withResponsive } from './with-responsive';
+export type { ResponsiveConfig } from './with-responsive';
