+++ projects/js-packages/charts/src/components/with-responsive/test/with-responsive.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import { withResponsive } from '../with-responsive';
+import { withResponsive } from '../index';
 import type { BaseChartProps } from '../../../types';
 
 // Mock the useParentSize hook
