+++ projects/js-packages/charts/src/components/with-responsive/with-responsive.tsx
